test: restructure unit tests
